Refactor import of render to AleliRenderer
